voice assistant/backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-multipart
python-dotenv
pydantic
pydantic-settings
passlib
bcrypt
python-jose[cryptography]
openai
langchain
langchain-openai
langchain-core
langchain-community
langchain-chroma
langchain-text-splitters
pypdf
chromadb
numpy
pypdf
scipy
alembic
opencv-python
websockets
httpx